feat(Modal)!: remove top-level sub-component(s)
