--- src/components/ModalBody/ModalBody.module.css
@@ -1,23 +1,0 @@
-@import '../../design-tokens/mixins.css';
-
-/*------------------------------------*\
-    # MODAL BODY
-\*------------------------------------*/
-
-/**
- * The body of the modal
- */
-.modal-body {
-  padding: var(--eds-size-4) var(--modal-horizontal-padding);
-  flex: 1;
-}
-
-.modal-body:focus-visible {
-  @mixin focusInside;
-}
-
-@supports not selector(:focus-visible) {
-  .modal-body:focus {
-    @mixin focusInside;
-  }
-}
