--- src/components/ModalBody/index.ts
@@ -1,1 +1,0 @@
-export { ModalBody as default } from './ModalBody';
