--- src/components/ModalFooter/ModalFooter.module.css
@@ -1,31 +1,0 @@
-/*------------------------------------*\
-    # MODAL FOOTER
-\*------------------------------------*/
-
-/**
- * Footer for the modal.
- */
-.modal-footer {
-  padding-top: var(--eds-size-2-and-half);
-  padding-bottom: var(--eds-size-2-and-half);
-  padding-left: var(--modal-horizontal-padding);
-  padding-right: var(--eds-size-3);
-
-  width: 100%;
-  background-color: var(--eds-theme-color-background-neutral-default);
-
-  border-bottom-left-radius: var(--eds-border-radius-lg);
-  border-bottom-right-radius: var(--eds-border-radius-lg);
-
-  z-index: var(--eds-z-index-100);
-}
-
-/**
- * Sticky variant of the footer.
- * Used when the modal variant is scrollable.
- */
-.modal-footer--sticky {
-  position: sticky;
-  bottom: 0;
-  box-shadow: var(--eds-box-shadow-xl);
-}
