--- src/components/ModalFooter/index.ts
@@ -1,1 +1,0 @@
-export { ModalFooter as default } from './ModalFooter';
