--- src/components/ModalHeader/index.ts
@@ -1,1 +1,0 @@
-export { ModalHeader as default } from './ModalHeader';
