--- src/components/ModalStepper/ModalStepper.module.css
@@ -1,12 +1,0 @@
-/*------------------------------------*\
-    # MODAL STEPPER
-\*------------------------------------*/
-
-/**
- * Stepper that resides in the modal footer.
- */
-.modal-stepper {
-  color: var(--eds-theme-color-icon-neutral-default);
-  display: flex;
-  gap: var(--eds-size-1-and-half);
-}
