--- src/components/ModalStepper/index.ts
@@ -1,1 +1,0 @@
-export { ModalStepper as default } from './ModalStepper';
