--- src/components/ModalTitle/index.ts
@@ -1,1 +1,0 @@
-export { ModalTitle as default } from './ModalTitle';
