+++ src/index.ts
@@ -54,11 +54,6 @@ export { default as Link } from './components/Link';
 export { default as LoadingIndicator } from './components/LoadingIndicator';
 export { default as Menu } from './components/Menu';
 export { default as Modal } from './components/Modal';
-export { default as ModalBody } from './components/ModalBody';
-export { default as ModalFooter } from './components/ModalFooter';
-export { default as ModalHeader } from './components/ModalHeader';
-export { default as ModalStepper } from './components/ModalStepper';
-export { default as ModalTitle } from './components/ModalTitle';
 export { default as NumberIcon } from './components/NumberIcon';
 export { default as PageHeader } from './components/PageHeader';
 export { default as PageLevelBanner } from './components/PageLevelBanner';
